Adding task

+++ api/index.php
@@ -18,42 +18,42 @@ function getAllTasks($pdo) {
     return $result;
 }
 
-switch ($_GET['action']) {
-    case 'add':
-        $name = $_GET['name'];
-        $statement = $pdo->prepare('INSERT INTO tasks (name) VALUES (:name)');
+if ( isset($_GET['action']) && $_GET['action'] == "getAll" ) {
+    $result = getAllTasks($pdo);
+    echo json_encode($result);
+} else {
+    switch ($_POST['action']) {
+        case 'add':
+            $name = $_POST['name'];
+            $statement = $pdo->prepare('INSERT INTO tasks (name) VALUES (:name)');
 
-        $statement->execute(['name' => $name]);
+            $statement->execute(['name' => $name]);
 
-        $result = getAllTasks($pdo);
-        echo json_encode($result);
-        break;
+            $result = getAllTasks($pdo);
+            echo json_encode($result);
+            break;
 
-    case 'edit':
-        $id = $_GET['id'];
-        $name = $_GET['name'];
-        $is_completed = $_GET['is_completed'];
+        case 'edit':
+            $id = $_GET['id'];
+            $name = $_GET['name'];
+            $is_completed = $_GET['is_completed'];
 
-        $sql = "UPDATE users SET name=?, is_completed=? WHERE id=?";
-        $stmt= $pdo->prepare($sql);
-        $stmt->execute([$name, $is_completed, $id]);
+            $sql = "UPDATE users SET name=?, is_completed=? WHERE id=?";
+            $stmt= $pdo->prepare($sql);
+            $stmt->execute([$name, $is_completed, $id]);
 
-        $result = getAllTasks($pdo);
-        echo json_encode($result);
-        break;
+            $result = getAllTasks($pdo);
+            echo json_encode($result);
+            break;
 
-    case 'delete':
-        $id = $_GET['id'];
+        case 'delete':
+            $id = $_GET['id'];
 
-        $q = $pdo->prepare("DELETE FROM tasks WHERE id = ?");
-        $q->execute(array($id));
+            $q = $pdo->prepare("DELETE FROM tasks WHERE id = ?");
+            $q->execute(array($id));
 
-        $result = getAllTasks($pdo);
-        echo json_encode($result);
-        break;
-
-    case 'getAll':
-        $result = getAllTasks($pdo);
-        echo json_encode($result);
-        break;
+            $result = getAllTasks($pdo);
+            echo json_encode($result);
+            break;
+    }
 } ?>
